refactor: 인증 상세 인증 투표 컴포넌트로 변경
